Revert "HTTPS on local ENV & support for production usage"
This reverts commit 9b8471e799b7e345fadf5ff510b2c926c4999f3e.
