+++ server/index.js
@@ -1,16 +1,11 @@
 import 'babel-polyfill';
+
 import config from 'config';
 import express from 'express';
+import http from 'http';
 
 import bootstrap from './bootstrap';
-import {log, normalizePort, startServer} from './utils';
-
-
-const fs = require('fs');
-const options = {
-    key: fs.existsSync('server.key') ? fs.readFileSync('server.key') : fs.readFileSync('key.pem'),
-    cert:fs.existsSync('server.crt') ? fs.readFileSync('server.crt') : fs.readFileSync('cert.pem')
-};
+import {log, normalizePort} from './utils';
 
 const app = express();
 
@@ -19,14 +14,25 @@ app.start = async () => {
     log.info('Starting Server...');
 
     const port = normalizePort(config.get('port'));
-    const securePort = normalizePort(config.get('securePort'));
 
     app.set('port', port);
 
     bootstrap(app);
 
-    startServer(app, securePort, options);
-    startServer(app, port);
+    const server = http.createServer(app);
+
+    server.on('error', (error) => {
+        if (error.syscall !== 'listen') throw error;
+        log.error(`Failed to start server: ${error}`);
+        process.exit(1);
+    });
+
+    server.on('listening', () => {
+        const address = server.address();
+        log.info(`Server listening ${address.address}:${address.port}`);
+    });
+
+    server.listen(port);
 };
 
 app.start().catch((err) => {
